IGNITE-13805 Split tests into core and indexing.
